Update patch set 1
Patch Set 1:

'No patches found' is also the result of having a patch with 'DROP-IN-RPM',
which is simple example of CR that does not create a new patch file.
Still, we do not want to fail the execution and we want to continue as this is not a fatal
message from our point of view.

Patch-set: 1