Reinstate: Avoid failure on 'No patches changed'
In an incoming CR does not produce a patch, we do not want to
fail the build, most common example being  DROP-IN-RPM changes.

rdopkg does return an error code when it encounters this case.
Now we look for this message and ignored failure if found.

While owners of rdopkg had reasons to consider this a failure
case, for downstream CI , this is not a valid failure reason.

We can ask rdopkg to change behaviour or to implement a
feature that ignores this case but we cannot wait for this to happen.

I avoided the case of looking specially for DROP-IN-RPM in the
last commit in order to make the code future proof, as other
magic tokens could get similar behaviour and there is no need
to replicate in Ansible same logic from rdopkg (is likely
to create more bugs in the future).

This reverts commit 607c53e7cd9bc5189854d81602639bf454fb3dd6.

Change-Id: I58fd4a4b8c1e12c101b71d4696f7b00c92c84c0d