WIP: add support to build patched RPMs on private component/dist-git branches
Sometimes there's a need to patch and build and RPM for a patch of
a component that's not aimed at a default/official rhos-XY... branch.

Currently I'm using this code to run by hand (with a few additional manual
steps) to help OpenDaylight devs test their patches before proposing
them to official rhos-XY branches or (more commonly) to test patches
to different odl-* components at once (i.e.: if some changes to odl-genius
rely on changes to odl-netvirt etc. and have to be tested together).

Please review the idea/code and comment whether that's something anyone
else would be interested in being supported.

Change-Id: I3dacff3ad84042ac59794c4f9a1a1aa31d0721e9