Conditionally copy RPMs locally and to undercloud
Before this change, in every execution we copied RPMs from patcher
node to localhost and then to undercloud.

In a case where tester=undercloud we don't need to do it because
if we do, we basically copy RPMs from undercloud to localhost and
back again to the undercloud. This whole process is redundant.

Change-Id: I95922134d638910e712a4fc5cfae3150ef6bca67