Install rdopkg from EPEL on el8 as well
The version from pip is not needed when a package is available,
and the rdopkg.el8 RPM has been available in EPEL8 for a while.

This means that EPEL repositories should be enabled on both
el7 and el8.

Moreover, at this time the version from pip requires a pyopenssl
version which can lead to errors in some cases, so the code
should be changed anyway to at least install rdopkg in a venv.

Change-Id: I52a863a0929e9b1b4015c608c063eb917d2bd272