Cleanup repositories handling
Use the shared rhos-release role to deploy rhos-release.
Use some generic links to refer to the internal rhpkg repositories.
Reorder the code which handles those repositories and consolidate
the operations for the various versions. The usage of shared
code and generic links reduces the distro-specific code
and will help adding RHEL9 support.

Change-Id: If6becf1c2c6b7030fe726da1f3d5c938644dccdd